Extract class file
